Refactor toggle-languange functionality
Add option to toggle the languange to
header and footer.
