--- constants/locale.ts
@@ -1,9 +1,0 @@
-export enum Locale {
-    English = "en",
-    German = "de"
-}
-
-export const LocaleIsoMap = {
-    [Locale.English]: "en-US",
-    [Locale.German]: "de-DE"
-};
