+++ custom.dic
@@ -5,8 +5,10 @@ coderabbit
 composables
 dropfile
 eavatar
+emojione
 ffpe
 firestore
+germany
 gserviceaccount
 hellmodus
 idat
